SEGMENTOR with Feature(50%) 1. runable under CentOS , G++4.8 2. There must be some BUGs; because currently we only get less than 50% F1 score and it doesn't increasing as traing . So Bad !

+++ src/segmentor/cws_module/lexicon_feature.cpp
@@ -50,7 +50,7 @@ void LexiconFeature::build_lexicon()
     {
         if( iter->second >= freq_threshold )
         {
-            size_t nr_utf8_char = UTF8Processing::utf8_char_len(iter->first);
+            unsigned nr_utf8_char = UTF8Processing::utf8_char_len(iter->first);
             if( nr_utf8_char > 1)
             {
                 // at least 2 word
